+++ src/DatabaseBackupServiceProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace AmdadulHaq\DatabaseBackup;
 
 use AmdadulHaq\DatabaseBackup\Console\BackupDatabaseCommand;
+use AmdadulHaq\DatabaseBackup\Console\RestoreDatabaseCommand;
 use Illuminate\Support\ServiceProvider;
 
 class DatabaseBackupServiceProvider extends ServiceProvider
@@ -23,7 +24,7 @@ class DatabaseBackupServiceProvider extends ServiceProvider
                 __DIR__.'/../config/database-backup.php' => config_path('database-backup.php'),
             ], 'database-backup-config');
 
-            $this->commands([BackupDatabaseCommand::class]);
+            $this->commands([BackupDatabaseCommand::class, RestoreDatabaseCommand::class]);
         }
     }
 }