+++ src/Exceptions/RestoreFailedException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AmdadulHaq\DatabaseBackup\Exceptions;
+
+use RuntimeException;
+
+class RestoreFailedException extends RuntimeException {}